requirements.txt
Django==1.10.1
django-disqus==0.5
django-import-export==0.5.0
django-nocaptcha-recaptcha==0.0.19
django-suit==0.2.21
django-wysiwyg-redactor==0.4.9.1
mock==2.0.0
pbr==1.10.0
Pillow==3.3.1
psycopg2==2.7.1
pytz==2016.6.1
six==1.10.0
tablib==0.11.5
PyGithub
django-mptt
